Add files via upload
I have created the front end of the signup and login page and could not manage to successfully connect two front ends together. An extensive amount of back end work is required in order to make a fully functional link as not only does the sign up page not show whether the sign up was successful or not, none of the buttons are corresponding. Added a live feature that keeps track of time until the next race, qualifier, sprint or grand prix. Moreover, I extended the team listings and their points in the championship, as well as driver points.